refactor urls
