+++ epilepsy12/views/management_views.py
