+++ epilepsy12/views/syndrome_views.py
